(fix): resolve issues with adding and removing project members

+++ app/src/main/java/com/ktun/ailabapp/data/remote/dto/response/LeaderBoardResponse.kt
@@ -6,7 +6,7 @@ data class LeaderboardUserResponse(
     val fullName: String,
 
     @SerializedName("totalScore")
-    val totalScore: Int,
+    val totalScore: Double,
 
     @SerializedName("profileImageUrl")
     val profileImageUrl: String?
